Increasing size of broker message count variable
For very large co-sims (in this case, mini_federate testing) the total message count was exceeding the max int value of 2.1 billion; the test in question should generate 6 billion messages. Increasing to a long gives us plenty of room.

+++ src/broker.cpp
@@ -101,7 +101,7 @@ int main(int argc, char **argv)
     static vector<int> vec_grant_time_calc;
     static vector<int> vec_granted_time;
     static vector<int> vec_granted_num;
-    static int publication_count = 0;
+    static long publication_count = 0;
 #endif
     
     fncs::start_logging();
